version 0.0.2
Major rewrite in this version. Since the aim is to write tab2xl2 so it
generates excel files almost ready for publishing, I decided to changed
the table presentation. Now column or row percentages are shown side by
side of frequencies, as opposed to under it. Also corrected a bug
pointed by Bryan.